requirements.txt
crispy-bootstrap4==2024.10
django==5.1.4
django-crispy-forms==2.3
django-extensions==3.2.3
django-filter==24.3
django_otp-1.5.4
django-tables2==2.7.5
